Flattens structure.
