+++ Sources/SweetBool/Predicates/PredicateTypeImplementations.swift
@@ -131,4 +131,3 @@ extension PredicateType {
         other
     }
 }
-
